Improve admonition preprocessing.
Multi-line admonitions couldn't be distinguished in the jupyter notebooks, as one couldn't tell where the admonition ends. Right now the preprocessor interpreter a blockquote inside an admonition as the visual clue used to fix that issue, and thus removes it so that it doesn't show up in the html output.