Fix #24
The preprocessing script parses the HTML admonitions inserted, but didn't take into account the fact that the admonition generated, which is delimited by backticks, might need to have its depth adjusted in case it already contains other backtick-delimited regions.
In #24, the remark that was badly formatted contained a piece of multiline code, delimited by `, which was clashing with the fact that the admonition is also defined by `.

Now, the code looks for backticks inside the admonition and picks a number of backticks that is enough to make sure there are no nesting issues.