Fix #21.
The INDEX ERRORs that were happening arose because the MyPreciousWs that was being loaded in the section about workspaces had been saved, somehow, with ⎕IO ← 0, which meant that loading it was throwing off the remainder of the calculations.

Supposedly, this issue would never really arise because the workspace was first being saved from within the notebook, and only then loaded. However, because the workspace already existed in the book folder, the first attempt to )save it was failing: the current workspace was the clear workspace and we were trying to save on top of a named workspace.

To fix this, I had to delete the MyPreciousWs from the book folder and re-execute the whole notebook, to both save a fresh copy of the workspace & to display the correct save message next to the )save command.
This commit includes a couple of immaterial changes to code cells which were used to force the re-execution of the (cached) notebook, a workaround that will hopefull be addressed with executablebooks/jupyter-book#1633.

Thoughts for the future include adding a preprocessing step where the MyPreciousWs is always deleted from the book folder (to make sure that the command )save always works) or a hidden cell in the notebook that deletes the workspace before saving.